requirements.txt
Flask==2.1.1  
mysql-connector-python==8.0.27  
flask-cors==3.0.10